+++ code/espurna/curtain_kingart.h
@@ -8,5 +8,5 @@ Copyright (C) 2019 by Albert Weterings
 */
 
 void kingartCurtainSetup();
-void curtainSetPosition(unsigned char id, long value);
+void curtainSetPosition(unsigned char id, int value);
 unsigned char curtainCount();
